requirements.txt
Django==4.1.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.3.6
pymongo==3.12.3
sqlparse==0.2.4
python-dotenv==1.0.0
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9 